# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.HealthResponse:
    properties:
      status:
        example: ok
        type: string
      timestamp:
        example: "2026-03-27T10:30:00Z"
        type: string
    type: object
  models.ErrorResponse:
    properties:
      message:
        example: internal server error
        type: string
      status:
        example: 500
        type: integer
    type: object
  models.Image:
    properties:
      createdAt:
        type: string
      filename:
        type: string
      height:
        type: integer
      id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
  models.PaginatedResponse:
    properties:
      hasMore:
        type: boolean
      images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      page:
        type: integer
      perPage:
        type: integer
      total:
        type: integer
    type: object
host: localhost:4000
info:
  contact: {}
  description: Snapteil backend API
  title: Snapteil API
  version: "1.0"
paths:
  /api/v1/health:
    get:
      description: Returns server health status with a timestamp
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HealthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Health check
      tags:
      - Health
  /api/v1/images:
    get:
      description: Returns a paginated list of images with optional tag filtering
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 5, max: 20)'
        in: query
        name: perPage
        type: integer
      - description: Comma-separated tags to filter by
        in: query
        name: tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get paginated images
      tags:
      - Image
  /api/v1/images/init:
    get:
      description: Returns the list of initial images with their metadata
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Image'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get initial images
      tags:
      - Image
  /api/v1/uploads:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a new image with title and tags
      parameters:
      - description: Image title
        in: formData
        name: title
        required: true
        type: string
      - description: Comma-separated tags
        in: formData
        name: tags
        type: string
      - description: Image file (JPEG, PNG, GIF, WebP, AVIF)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Upload an image
      tags:
      - Upload
swagger: "2.0"